Refactor code and add Docker support
This commit moves the 'create_file' functionality from 'views.py' to 'services.py' for better code organization and separation of concerns. Docker support is also added with the inclusion of 'Dockerfile' and 'docker-compose.yaml' files. The docker setup ensures that our app is containerized and can be run in any environment seamlessly. Additionally, the 'uvicorn' package is added to the requirements to aid in running our ASGI application. Relevant logs within the 'create' method have been also updated to improve error tracking.